+++ src/main/java/com/rutaji/exaqua/block/ModBlocks.java
@@ -19,7 +19,7 @@ import java.util.function.Supplier;
 public class ModBlocks {
     //region register
     public  static final DeferredRegister<Block> BLOCKS = DeferredRegister.create(ForgeRegistries.BLOCKS, ExAqua.MOD_ID);
-    //helper method to register block and its item
+
 
     private static <T extends Block>RegistryObject<T> registerBlock(String name, Supplier<T> block){
         RegistryObject<T> toReturn = BLOCKS.register(name,block);
@@ -29,6 +29,7 @@ public class ModBlocks {
     private static <T extends Block> void registerBlockItem(String name,RegistryObject<T> block){
         ModItems.ITEMS.register(name,()-> new BlockItem(block.get(),new Item.Properties().group(CustomItemGroup.EX_AQUA_GROUP)));
     }
+    //endregion
     public static RegistryObject<Block> GetSIEVE(SieveTiers t)
     {
         switch (t)
@@ -45,7 +46,7 @@ public class ModBlocks {
     public static void register(IEventBus eventBus){
         BLOCKS.register(eventBus);
     }
-    //endregion
+
     public static final RegistryObject<Block> IRONSIEVE =registerBlock("ironsieve",() -> new SieveBlock(AbstractBlock.Properties.create(Material.IRON).hardnessAndResistance(6), SieveTiers.iron));
     public static final RegistryObject<Block> GOLDSIEVE =registerBlock("goldsieve",() -> new SieveBlock(AbstractBlock.Properties.create(Material.IRON).hardnessAndResistance(6), SieveTiers.gold));
     public static final RegistryObject<Block> FROGIUMSIEVE =registerBlock("frogiumsieve",() -> new SieveBlock(AbstractBlock.Properties.create(Material.IRON).hardnessAndResistance(6), SieveTiers.frogium));
@@ -53,5 +54,6 @@ public class ModBlocks {
     public static final RegistryObject<Block> SQUEEZER =registerBlock("squeezer",() -> new squeezerBlock(AbstractBlock.Properties.create(Material.IRON).hardnessAndResistance(6)));
     public static final RegistryObject<Block> F_ORE_END =registerBlock("frogium_ore_end",() -> new Block(AbstractBlock.Properties.create(Material.ROCK).hardnessAndResistance(4)));
     public static final RegistryObject<Block> F_ORE_STONE =registerBlock("frogium_ore_stone",() -> new FrogiumOre(AbstractBlock.Properties.create(Material.ROCK).hardnessAndResistance(2)));
+    public static final RegistryObject<Block> F_BLOCK =registerBlock("frogium_block",() -> new FrogiumBlock(AbstractBlock.Properties.create(Material.IRON).hardnessAndResistance(4)));
 
 }
