+++ src/main/resources/assets/exaqua/blockstates/frogium_block.json
@@ -1,0 +1,5 @@
+{
+  "variants": {
+    "": {"model":"exaqua:block/frogium_block" }
+  }
+}
