+++ src/main/resources/assets/exaqua/models/block/frogium_block.json
@@ -1,0 +1,6 @@
+{
+  "parent": "block/cube_all",
+  "textures": {
+    "all": "exaqua:block/frogium_block"
+  }
+}
