+++ src/main/resources/assets/exaqua/models/item/frogium_block.json
@@ -1,0 +1,3 @@
+{
+  "parent": "exaqua:block/frogium_block"
+}
