+++ src/main/resources/data/exaqua/loot_tables/blocks/frogium_block.json
@@ -1,0 +1,13 @@
+{
+  "pools": [
+    {
+      "rolls": 1,
+      "entries": [
+        {
+          "type": "minecraft:item",
+          "name": "exaqua:frogium_block"
+        }
+      ]
+    }
+  ]
+}
