+++ src/main/resources/data/exaqua/loot_tables/blocks/squeezer.json
@@ -1,0 +1,13 @@
+{
+  "pools": [
+    {
+      "rolls": 1,
+      "entries": [
+        {
+          "type": "minecraft:item",
+          "name": "exaqua:squeezer"
+        }
+      ]
+    }
+  ]
+}
