+++ src/main/resources/data/exaqua/recipes/frogium_block.json
@@ -1,0 +1,15 @@
+{
+  "type": "minecraft:crafting_shaped",
+  "pattern": [
+    "##",
+    "##"
+  ],
+  "key": {
+    "#": {
+      "item": "exaqua:frogium"
+    }
+  },
+  "result": {
+    "item": "exaqua:frogium_block"
+  }
+}
