+++ src/main/resources/data/exaqua/recipes/frogium_ore.json
@@ -1,0 +1,18 @@
+{
+  "type": "minecraft:crafting_shaped",
+  "pattern": [
+    "#s",
+    "s#"
+  ],
+  "key": {
+    "#": {
+      "item": "exaqua:frogium_dust"
+    },
+    "s": {
+      "item": "minecraft:stone"
+    }
+  },
+  "result": {
+    "item": "exaqua:frogium_ore_stone"
+  }
+}
